+++ src/helpers.ts
@@ -1,16 +1,26 @@
-import { z } from "zod";
-import Robj from "./types";
+function isNumArray(array: unknown[]): array is number[] {
+  return typeof array[0] === "number";
+}
 
 function as_vector(x: number[]): Float64Array;
 function as_vector(x: string[]): string[] & { r_type: "string_array" };
 function as_vector(x: string[] | number[]) {
-  if (typeof x[0] === "number") {
-    return new Float64Array(x as number[]);
+  if (isNumArray(x)) {
+    return new Float64Array(x);
   }
   const res: string[] & { r_type: "string_array" } = x as any;
   res.r_type = "string_array";
   return res;
 }
+
+const promisify =
+  (func: Function) =>
+  (...args: any[]) =>
+    new Promise((resolve, reject) =>
+      func(...args, (err: Error, result: any) =>
+        err ? reject(err) : resolve(result)
+      )
+    );
 
 // type ObjectWithDim = {
 //   r_attributes: {
@@ -35,4 +45,4 @@ function as_vector(x: string[] | number[]) {
 //   }
 // }
 
-export { as_vector };
+export { as_vector, promisify };