+++ src/index.ts
@@ -135,5 +135,6 @@ const RserveClient = {
 };
 
 export default RserveClient;
+export { Robj };
 export * as Rfmt from "./helpers";
 export type * from "./types";